[feature] unit tests added and added/changed Controllers to work with Open API Specs

+++ src/main/java/com/ubs/m295_projectapplication/service/extractor/ProjectSetExtractor.java
@@ -23,7 +23,7 @@ public class ProjectSetExtractor implements ResultSetExtractor<List<Project>> {
         if(rs.next()) {
             Project project = new Project();
             project.setProjectName(rs.getString("projectName"));
-            project.setProjectId(rs.getLong("projectId"));
+            project.setProjectId(rs.getInt("projectId"));
             projectArrayList.add(project);
             return projectArrayList;
         }else{
